requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Background Jobs
celery==5.3.6
redis==5.0.1

# OpenAI
openai==1.10.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
aiofiles==23.2.1

# Testing
pytest==7.4.4
httpx==0.26.0
